test(codegen): consolidate generator tests into generate_test.py, align discovery pattern
- Merge TestDeprecatedField from test_generate.py into generate_test.py so all
  schema generator tests live in one file with a consistent *_test.py naming
  convention
- Remove test_generate.py (now superseded)
- Update generator_test.go to use *_test.py discovery pattern, matching the
  CI "Test schema generator" step exactly — previously test_generate.py was
  invisible to CI and generate_test.py (from #183) was invisible to the Go test
